+++ crates/cubed-application/src/ports/mod.rs
@@ -1,3 +1,4 @@
+pub mod console;
 pub mod downloader;
 pub mod file_system;
 pub mod java_manager;
@@ -5,6 +6,7 @@ pub mod port_manager;
 pub mod process_manager;
 pub mod server_repository;
 
+pub use console::{ConsoleLine, ConsoleCallback, ConsoleManager};
 pub use downloader::{DownloadedJar, Downloader};
 pub use file_system::FileSystemManager;
 pub use java_manager::{JavaInstallation, JavaManager};